Top level prepend() and adjust() methods

+++ src/irvine/oeis/a035/A035401.java
@@ -10,7 +10,7 @@ import irvine.oeis.Sequence1;
  */
 public class A035401 extends Sequence1 {
 
-  private int mM = 0;
+  private long mM = 0;
   private final IntegerPartition mPart;
 
   protected A035401(final int residue) {
